Add project about transmission quality and Raspberry Pi deployment
* Update 2022-01-25-Related-Projects.md

I have now posted this project here. It was previously in posts #1084 and #1085 and I was asked to move it here.

* Update 2022-01-25-Related-Projects.md

I have incorporated the changes that ann0see suggested.

* Update 2022-01-25-Related-Projects.md

I have added the following sentence:
The project was developed by students of electrical engineering and computer science at the FH Aachen University of Applied Sciences.

* Update 2022-01-25-Related-Projects.md

I have shortened the headline